fix new clippy warnings

+++ primitives/src/circuit/signature/schnorr.rs
@@ -186,7 +186,7 @@ where
         chal_input.extend(msg);
 
         let challenge = self.rescue_sponge_with_padding(&chal_input, 1)?[0];
-        let c_bits = self.unpack(challenge, field_bit_len::<F>() as usize)?;
+        let c_bits = self.unpack(challenge, field_bit_len::<F>())?;
         Ok(c_bits[..challenge_bit_len::<F>()].to_vec())
     }
 }
